+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Removed `exists()` on model class.
 - Removed `title` setting on model properties.
 - Removed `ACLModel::setRequester()` and `ACLModel::getRequester()`
+- Removed `Errors::errors()`, `Errors::messages()`, and `Errors::push()`
 
 ## 0.9.1 - 2019-11-16
 ## Fixed